services/api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
boto3>=1.35.0
python-magic>=0.4.27
Pillow>=11.0.0
PyPDF2>=3.0.0

# RAG / Vector / LLM
lancedb~=0.20.0
sentence-transformers>=3.0.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
tiktoken>=0.8.0
pyarrow>=18.0.0

# Dev / quality
ruff>=0.8.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
